requirements.txt
streamlit==1.31.1
langchain==0.1.3
langchain-community==0.0.15
langchain-core==0.1.15
langchain-google-genai==0.0.6
python-dotenv==1.0.0
chromadb==0.4.22
faiss-cpu==1.7.4
beautifulsoup4==4.12.2
pypdf==3.12.2